Meeting planner finished
The meeting planner is as good as finished; excepting the 'known issues', not least of which is that we don't know for sure in which format de data is stored in the database and how certain decisions are made based on that data. Please see github issues page for more information.

New in this version:
* Using icons in detail view
* Colors are no longer chosen randomly but actually indicate the availability of people who responded for that time slot
* Setting preference and comment for a slot now really works, rather than setting incorrect or incomplete information
* The model is updated instantly after the user makes a change, giving real time feedback
* The code was cleaned thoroughly. Code was split into functions, documented, redundant code was removed, the modal view has been split to a seperate file and the overall structure is now more clear