nav-bar-title now works correctly.
When a new view is loaded, the current slide index is requested and for this index, the name of this slide is looked up in app.js. This name is then set as the title for the nav bar.
The view.enter event is necessary, because the title is set after the controller is instantiated by the nav bar, meaning that any title we set will be overridden by the nav bar.
THe disadvantage of this method is that sometimes the name of the last slide is shown for a fraction of a second before the correct name appears. As far as I can find out there is nothing to do about that except get rid of slides and move to tabs.